Refactoring the parakeet
